pyproject.toml
[project]
name = "saphi-custom-tracks"
version = '1.0.0'
description = "Custom Track Repository for Project Saphi"
dependencies = [
    "flask",
    "flask-migrate",
    "flask-sqlalchemy",
    "flask-jwt-extended",
    "marshmallow",
    "psycopg2-binary"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
